sandbox - move some files to electron-sandbox that are ready
