+++ src/vs/base/node/ports.ts
@@ -6,15 +6,6 @@
 import * as net from 'net';
 
 /**
- * @returns Returns a random port between 1025 and 65535.
- */
-export function randomPort(): number {
-	const min = 1025;
-	const max = 65535;
-	return min + Math.floor((max - min) * Math.random());
-}
-
-/**
  * Given a start point and a max number of retries, will find a port that
  * is openable. Will return 0 in case no free port can be found.
  */
