+++ src/vs/workbench/contrib/extensions/electron-sandbox/extensionsSlowActions.ts
