+++ src/vs/workbench/contrib/extensions/electron-sandbox/reportExtensionIssueAction.ts
